(tests) #16 Moved tests
All tests where moved from BinaryFunctionsTest to BinaryFileFunctionsTest.
Creating a new file seemed to cause licensing errors.
